tenporary change to unbreak tests

+++ nifty6/sugar.py
@@ -498,7 +498,8 @@ def calculate_position(operator, output):
     else:
         cov = 1e-3*output.val.max()**2
     invcov = ScalingOperator(output.domain, cov).inverse
-    d = output + invcov.draw_sample(from_inverse=True)
+    # FIXME!!!
+    d = output + invcov.draw_sample(dtype=np.float64, from_inverse=True)
     lh = GaussianEnergy(d, invcov) @ operator
     H = StandardHamiltonian(
         lh, ic_samp=GradientNormController(iteration_limit=200))
